+++ bolhacparada.h
@@ -1,6 +1,6 @@
 #ifndef BOLHACPARADA_H
 #define BOLHACPARADA_H
 
-void bubbleSortParada (int* vet, int tam, int* troca, int* comp);
+void bubbleSortParada (int* vet, int tam, long* troca, long* comp);
 
 #endif
