+++ quicksortmeio.h
@@ -1,6 +1,6 @@
 #ifndef QUICKSORTMEIO_H
 #define QUICKSORTMEIO_H
 
-void quickSortMeio (int* vet, int esq, int dir, int* comp, int* troca);
+void quickSortMeio (int* vet, int esq, int dir, long* comp, long* troca);
 
 #endif
